finetune_demo/scripts/requirements.txt
absl-py==2.1.0
accelerate==1.1.1
aiobotocore==2.15.2
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aioitertools==0.12.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.13.1
beautifulsoup4==4.12.2
bitsandbytes==0.44.1
bleach==6.1.0
blinker==1.4
botocore==1.35.36
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.7
comm==0.2.0
contourpy==1.3.0
cryptography==3.4.8
cycler==0.12.1
datasets==2.20.0
dbus-python==1.2.18
debugpy==1.8.0
decorator==5.1.1
deepspeed==0.14.4
defusedxml==0.7.1
dill==0.3.8
distro==1.7.0
einops==0.8.0
entrypoints==0.4
exceptiongroup==1.1.3
executing==2.0.1
faiss-cpu==1.9.0
fastapi==0.115.4
fastjsonschema==2.18.1
ffmpy==0.4.0
filelock==3.9.0
fonttools==4.54.1
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.10.0
fuzzywuzzy==0.18.0
gradio==4.44.1
gradio_client==1.3.0
grpcio==1.67.1
gunicorn==23.0.0
h11==0.14.0
hjson==3.1.0
httpcore==1.0.6
httplib2==0.20.2
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.4
importlib-metadata==4.6.4
importlib_resources==6.4.5
ipykernel==6.26.0
ipython==8.17.2
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
jeepney==0.7.1
jieba==0.42.1
Jinja2==3.1.4
jiter==0.7.0
jmespath==1.0.1
joblib==1.4.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter-archive==3.4.0
jupyter-contrib-core==0.4.2
jupyter-contrib-nbextensions==0.7.0
jupyter-events==0.9.0
jupyter-highlight-selected-word==0.2.0
jupyter-lsp==2.2.0
jupyter-nbextensions-configurator==0.6.3
jupyter_client==7.4.9
jupyter_core==5.5.0
jupyter_server==2.10.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.8
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
keyring==23.5.0
kiwisolver==1.4.7
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
Levenshtein==0.26.1
lxml==4.9.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.9.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
nbclassic==1.0.0
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.0
ninja==1.11.1.1
nltk==3.8.1
notebook==6.5.5
notebook_shim==0.2.3
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.0
openai==1.54.3
orjson==3.10.11
overrides==7.4.0
packaging==23.2
pandas==2.2.3
pandocfilters==1.5.0
parso==0.8.3
peft==0.13.2
pexpect==4.8.0
pillow==10.4.0
platformdirs==3.11.0
prometheus-client==0.18.0
prompt-toolkit==3.0.39
propcache==0.2.0
protobuf==5.28.3
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==18.0.0
pyarrow-hotfix==0.6
pycparser==2.21
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.16.1
PyGObject==3.42.1
PyJWT==2.3.0
pyngrok==7.2.1
pyparsing==2.4.7
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-json-logger==2.0.7
python-Levenshtein==0.26.1
python-multipart==0.0.17
pytz==2024.2
PyYAML==6.0.1
pyzmq==24.0.1
RapidFuzz==3.10.1
referencing==0.30.2
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rouge-chinese==1.0.3
rpds-py==0.12.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
ruff==0.7.3
s3fs==2024.10.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
SecretStorage==3.3.1
semantic-version==2.10.0
Send2Trash==1.8.2
sentence-transformers==3.2.1
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sse-starlette==2.1.3
stack-data==0.6.3
starlette==0.41.2
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
terminado==0.17.1
threadpoolctl==3.5.0
tiktoken==0.8.0
timm==1.0.11
tinycss2==1.2.1
tokenizers==0.19.1
tomli==2.0.1
tomlkit==0.12.0
torch==2.5.1
torchaudio==2.1.0+cu118
torchvision==0.20.1
tornado==6.3.3
tqdm==4.67.0
traitlets==5.13.0
transformers==4.44.2
triton==3.1.0
typer==0.13.0
types-python-dateutil==2.8.19.14
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
urllib3==2.2.3
uvicorn==0.32.0
wadllib==1.3.6
wcwidth==0.2.9
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
websockets==12.0
Werkzeug==3.1.3
widgetsnbextension==4.0.9
wrapt==1.16.0
xxhash==3.5.0
yarl==1.17.1
zipp==1.0.0
